revert: revert age col back to int type

+++ scrape_and_score/sql/create_player_demographics_table.sql
@@ -3,7 +3,7 @@
 CREATE TABLE player_demographics (
 	player_id INT NOT NULL, 
 	season INT NOT NULL, 
-	age FLOAT, 
+	age INT, 
 	height FLOAT,
 	weight FLOAT,
 	PRIMARY KEY (player_id, season),
